requirements.txt
alembic==1.4.3
aniso8601==8.1.0
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.4
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
distlib==0.3.1
docopt==0.6.2
ecdsa==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.5.3
Flask-Moment==0.11.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
pipreqs==0.4.10
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycodestyle==2.6.0
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
python-dateutil==2.6.0
python-editor==1.0.4
python-jose==3.2.0
pytz==2020.5
requests==2.25.1
rsa==4.7
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
urllib3==1.26.2
virtualenv==20.2.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
yarg==0.1.9
